pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app", "config"]

[project]
name = "cherrypy-base"
version = "0.1.0"
description = "CherryPy application skeleton"
requires-python = ">=3.11"
dependencies = [
    "cherrypy>=18.10.0",
    "jinja2>=3.1.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "ruff>=0.8.0",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
